requirements.txt
nvidia-ml-py3==7.352.0
transformers
sentence-transformers==2.2.2
accelerate==0.25.0
wget==3.2
scikit-learn==1.1.3
nltk==3.8.1
stqdm==0.0.4
datasets==2.6.1
python-dotenv
einops
torch
langchain
huggingface_hub
text_generation
xmltodict
langchainhub
spacy
scispacy